pyproject.toml
[tool.poetry]
name = "docker-ready-cli"
version = "0.1.0"
description = ""

authors = ["impocode <impocode@impocode.one>"]
maintainers = ["impocode <impocode@impocode.one>"]

license= "MIT"
readme = "readme.md"

keywords = ["docker"]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

homepage = "https://github.com/docker-ready/cli"
repository = "https://github.com/docker-ready/cli"
documentation = "https://github.com/docker-ready/cli"

[tool.poetry.dependencies]
python = "^3.12"
click = "^8.1.7"
rich = "^13.7.1"
docker-ready = "^0.2.0"
python-dateutil = "^2.9.0.post0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.9"
pytest = "^8.2.1"

[tool.poetry.scripts]
docker-ready = "docker_ready_cli:cli"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = [
    "F",  # pyflakes
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "I",  # isort
    "UP", # pyupgrade
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
